// src/app/models/User.model.ts
export class User {
    id: number;
    email: string;
    userName: string;
    montantDepose: number;
    btc: number;
    eth: number;
    bnb: number;
    ltc: number;
    eos: number;
    bch: number;
    trx: number;
    neo: number;
    ada: number;
    xrp: number
    // id: number;
    // email: string;
    // username: string;
    // montantDepose: number;
    // acessToken: string;
    // tokenType: string;
    // ada: number;
    // bch: number;
    // bnb: number;
    // btc: number;
    // eos: number;
    // eth: number;
    // ltc: number;
    // neo: number;
    // trx: number;
    // xrp: number
  }
  